+++ src/configurator/knobs.ts
@@ -119,7 +119,7 @@ export const knobs = [
     id: 'editorBackground',
     label: 'Background',
     section: 'code editor',
-    cssProperty: '--playground-file-editor-background-color',
+    cssProperty: '--playground-file-editor-background',
     default: '',
     unsetLabel: 'From theme',
   }),
@@ -128,7 +128,7 @@ export const knobs = [
   color({
     id: 'filePickerBackground',
     label: 'Background',
-    cssProperty: '--playground-file-picker-background-color',
+    cssProperty: '--playground-file-picker-background',
     default: '#ffffff',
     section: 'file picker',
   }),
@@ -144,7 +144,7 @@ export const knobs = [
   color({
     id: 'previewToolbarBackground',
     label: 'Background',
-    cssProperty: '--playground-preview-toolbar-background-color',
+    cssProperty: '--playground-preview-toolbar-background',
     default: '#ffffff',
     section: 'preview toolbar',
   }),